Improved code base
* Fixed docstrings
* Code cleanup
  - Removed unused subprocess module
  - Fixed variable names (e -> err)
* Moved populate_properties() and parsecli()
  - __init__.py -> cli.py
  - Adapted testcases
* Moved localname(), validroots; adapted namespace
  - Moved localname() and validroots: xmlhandler.py -> xmlutil.py
  - Imported NS from docmanager.core and used it instead of self.__namespace
